fix(commands): promote stripMarkdownFences to core/markdown.ts, fix reflect.ts <think> stripping
Fixes #319
